Updated PezUI-LoginScripts to addresses an issue with VuhDo
For some reason, VuhDo does not auto-load the best profile (based on
group size) after login or loading screen. This causes issues if the
group size changes while logged out or in a loading screen.

To address this issue, a call to VuhDo's auto-load function is made in
response to the PLAYER_ENTERING_WORLD event.