Created Shortcut action bar and tweaked panel layout
The shortcut bar is somewhat of a hack for now, until more Ion features
are released. The features used to achieve the shortcut action bar
design previously (with Macaroon) used button margins, which is not
available in the current Ion beta release. Once this feature is made
available in Ion, three shortcut bars will be merged into one.

In laying out the shortcut action bar the panel layout was adjusted,
requiring adjustment of panel sizes, and positioning of MSBT scroll
areas.